Migrate clustersState to jotai
