+++ package.json
@@ -87,6 +87,7 @@
     "i18next-http-backend": "^3.0.2",
     "immutable": "^5.1.1",
     "jotai": "^2.13.0",
+    "jotai-effect": "^2.0.5",
     "js-tiktoken": "^1.0.20",
     "js-yaml": "^4.1.0",
     "jsdom": "^26.1.0",
