+++ src/state/utils/effects.ts
@@ -1,11 +1,6 @@
 import { AtomEffect } from 'recoil';
-import { ClustersState } from '../clustersAtom';
-import { Cluster } from '../clusterAtom';
-import { ClusterStorage } from '../types';
 
 type LocalStorageEffectFn = <T>(localStorageKey: string) => AtomEffect<T>;
-
-const inMemoryClusters: ClustersState = {};
 
 export const localStorageEffect: LocalStorageEffectFn = localStorageKey => ({
   setSelf,
@@ -26,72 +21,4 @@ export const localStorageEffect: LocalStorageEffectFn = localStorageKey => ({
   onSet(newValue =>
     localStorage.setItem(localStorageKey, JSON.stringify(newValue)),
   );
-};
-
-type ClusterStorageEffectFn = <T>(localStorageKey: string) => AtomEffect<T>;
-
-export const clusterStorageEffect: ClusterStorageEffectFn = clusterStorageKey => ({
-  setSelf,
-  onSet,
-}) => {
-  setSelf(previousValue => {
-    const savedValue = {
-      ...JSON.parse(localStorage.getItem(clusterStorageKey) || '{}'),
-      ...JSON.parse(sessionStorage.getItem(clusterStorageKey) || '{}'),
-      ...inMemoryClusters,
-    };
-
-    try {
-      if (savedValue !== null) return savedValue;
-      return previousValue || {};
-    } catch (e) {
-      console.warn('Cannot get clusters', e);
-      return previousValue || {};
-    }
-  });
-
-  onSet(newValue => {
-    const localStorageClusters: ClustersState = {};
-    const sessionStorageClusters: ClustersState = {};
-    const clusters = newValue ? Object.entries(newValue) : [];
-
-    for (const key in clusters) {
-      const clusterName: string = clusters[key as keyof ClustersState][0];
-      let cluster = clusters[key as keyof ClustersState][1] as Cluster;
-      const storage: ClusterStorage = cluster.config
-        ? cluster.config.storage || 'sessionStorage'
-        : 'sessionStorage';
-
-      switch (storage) {
-        case 'localStorage':
-          localStorageClusters[clusterName] = cluster;
-          break;
-        case 'sessionStorage':
-          sessionStorageClusters[clusterName] = cluster;
-          break;
-        case 'inMemory':
-          inMemoryClusters[clusterName] = cluster;
-          break;
-        default:
-          console.warn(
-            clusterName,
-            ': Unknown storage type ',
-            cluster?.config?.storage,
-            'saving in localStorage',
-          );
-          cluster.config = { storage: 'localStorage' };
-          localStorageClusters[clusterName] = cluster;
-          break;
-      }
-    }
-
-    localStorage.setItem(
-      clusterStorageKey,
-      JSON.stringify(localStorageClusters),
-    );
-    sessionStorage.setItem(
-      clusterStorageKey,
-      JSON.stringify(sessionStorageClusters),
-    );
-  });
 };
