+++ src/state/utils/getClustersInfo.ts
@@ -3,6 +3,7 @@ import { useNavigate } from 'react-router';
 
 import { clusterState } from 'state/clusterAtom';
 import { clustersState } from 'state/clustersAtom';
+import { useAtom } from 'jotai';
 
 export type useClustersInfoType = ReturnType<typeof useClustersInfo>;
 
@@ -10,7 +11,7 @@ export function useClustersInfo() {
   const navigate = useNavigate();
 
   const [currentCluster, setCurrentCluster] = useRecoilState(clusterState);
-  const [clusters, setClusters] = useRecoilState(clustersState);
+  const [clusters, setClusters] = useAtom(clustersState);
 
   return {
     currentCluster,
